Use mock fetch in all tests
